Move MailStore classes into a separate package - for better organized code.
