+++ src/main/java/org/masukomi/aspirin/core/store/mail/MailStore.java
@@ -1,4 +1,4 @@
-package org.masukomi.aspirin.core.store;
+package org.masukomi.aspirin.core.store.mail;
 
 import javax.mail.internet.MimeMessage;
